implemeneted and tested filter preferences on search
Moved toString to function instead of directly in the preference in the repo
Implemented try catch on datastore access with log
changed uistate distance from int to double

+++ app/src/main/java/com/digitalarchitects/rmc_app/presentation/screens/search/SearchUIState.kt
@@ -6,7 +6,7 @@ data class SearchUIState(
     val date: String = "", // TODO: Make LocalDate, maybe with LocalDate.now()
     val location: String = "",
     val price: Double = 0.0,
-    val distance: Int = 0,
+    val distance: Double = 0.0,
     val engineTypeIce: Boolean = true,
     val engineTypeBev: Boolean = true,
     val engineTypeFcev: Boolean = true
